+++ app/src/main/java/com/ob/foodapp/FoodAppActivity.kt
@@ -2,12 +2,13 @@ package com.ob.foodapp
 
 import android.os.Bundle
 import android.os.PersistableBundle
-import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
+import com.google.firebase.auth.FirebaseAuth
 import com.ob.foodapp.databinding.ActivityFoodAppBinding
+import com.ob.foodapp.feature.signin.presentation.model.UiUser
 import org.kodein.di.Kodein
 import org.kodein.di.KodeinAware
 import org.kodein.di.KodeinTrigger
@@ -23,6 +24,8 @@ class FoodAppActivity : AppCompatActivity(), KodeinAware {
 
     @SuppressWarnings("LeakingThisInConstructor")
     override val kodeinContext = kcontext<AppCompatActivity>(this)
+
+    private val navigator = FoodAppNavigator()
 
     override val kodein: Kodein by retainedKodein {
         extend(parentKodein)
@@ -42,6 +45,28 @@ class FoodAppActivity : AppCompatActivity(), KodeinAware {
         super.onCreate(savedInstanceState)
         binding = ActivityFoodAppBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        // TODO: Use the AuthViewModel instance for this
+        checkSession()
+    }
+
+    private fun checkSession() {
+        val navController = findNavController(R.id.navHostFragment)
+        val currentUser = FirebaseAuth.getInstance().currentUser
+
+        if (currentUser == null) {
+            navigator.launchAppWithSignIn(navController)
+        } else {
+            // TODO: Implement in AuthViewModel for avoid use this Builder
+            navigator.launchAppWithResult(
+                navController,
+                UiUser(
+                    uid = currentUser.uid,
+                    name = currentUser.displayName ?: "",
+                    email = currentUser.email ?: ""
+                )
+            )
+        }
     }
 
     override fun onSupportNavigateUp(): Boolean {