+++ CHANGELOG.md
@@ -6,3 +6,11 @@
 
 * Implement Architecture for Android app
 * Defining libraries and configurations using Kotlin DSL
+* Implement Sign In, Result and Get Profile features
+
+## 0.2.0 (2021/08/22)
+* Migrate Firebase project
+* Fix Get Profile for Firestore
+* Implement SignOut
+* Implement Edit Profile
+* Fixes for Navigation